% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA1num.R
\name{EDA1num}
\alias{EDA1num}
\title{Exploratory Data Analysis from a cuantitative varaible}
\usage{
EDA1num(dat, var)
}
\arguments{
\item{dat}{A data frame}

\item{var}{A numeric varaible from the data frame}
}
\value{
The scalar product of \code{x} and \code{y}.
}
\description{
Exploratory Data Analysis from a cuantitative varaible (numeriacal a graphical solutions)
}
\examples{
EDA1num(airquality, Ozone)

}
